% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissoxygen_function_script.r
\name{dissoxygen_function}
\alias{dissoxygen_function}
\title{Classify Monthly Dissolved Oxygen Conditions}
\usage{
dissoxygen_function(dissoxygendata)
}
\arguments{
\item{dissoxygendata}{A data frame containing daily dissolved oxygen observations.
The data frame must include: date of observation and daily dissolved oxygen concentration in milligrams per liter (mg/L)}
}
\value{
A data frame summarizing monthly dissolved oxygen conditions, including the month, mean dissolved oxygen concentration (mg/L), and dissolved oxygen classification
}
\description{
Processes daily dissolved oxygen (DO) measurements from a lake or water body
Organizes the data by water year month order (October through September)
Calculates monthly mean DO concentrations, and assigns a classififcation to each month based on established oxygen thresholds

Hypoxic   = Very low oxygen conditions, DO < 2 mg/L
Stressful = Low oxygen conditions, DO 2-5 mg/L
Moderate  = Moderate oxygen conditions, DO 5-8 mg/L
Excellent = High oxygen conditions, DO > 8 mg/L
}
